fix: fix coinbase verify bug while restarting

+++ src/test/transaction_tests.cpp
@@ -20,6 +20,7 @@
 #include "taskcancellation.h"
 #include "test/jsonutil.h"
 #include "test/scriptflags.h"
+#include "uint256.h"
 #include "utilstrencodings.h"
 #include "validation.h" // For CheckRegularTransaction
 #include "chainparams.h"
@@ -107,7 +108,7 @@ void RunTests(Config& globalConfig, UniValue& tests, bool should_be_valid){
             CValidationState state;
 
             fValid = tx.IsCoinBase() 
-                                    ? CheckCoinbase(tx, state, MAX_TX_SIGOPS_COUNT_BEFORE_GENESIS, MAX_TX_SIZE_CONSENSUS_BEFORE_GENESIS, false)
+                                    ? CheckCoinbase(tx, state, MAX_TX_SIGOPS_COUNT_BEFORE_GENESIS, MAX_TX_SIZE_CONSENSUS_BEFORE_GENESIS, false, uint256())
                                     : CheckRegularTransaction(tx, state, MAX_TX_SIGOPS_COUNT_BEFORE_GENESIS, MAX_TX_SIZE_CONSENSUS_BEFORE_GENESIS, false);
 
             if(!(fValid && state.IsValid())) {
